QR Code Generator/requirements.txt
Pillow==9.4.0
pypng==0.20220715.0
PyQRCode==1.2.1
